Move Theme to contexts directory
